Locationの修正2
